requirements.txt
alembic==1.10.4
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
h11==0.14.0
httptools==0.5.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
pydantic==1.10.7
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.11
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.2
